test(cli): add comprehensive tests for repo and utility commands
- internal/cli/repo_command_test.go: Implement tests for repo commands including add, list, remove, and handling of various scenarios.
- internal/cli/root_command_test.go: Add tests for root command functionalities including version and help commands.
- internal/cli/test_helpers_test.go: Introduce helper functions for testing HTTP servers and path expectations.
- internal/cli/util_command_test.go: Create tests for utility commands, including digest and zstd compression functionalities.
